bitsplit: create shared unit
for capabilities shared between compression and decompression
