+++ src/openzl/codecs/bitSplit/encode_bitSplit_kernel.c
@@ -4,20 +4,6 @@
 
 #include <string.h>
 
-size_t ZS_bitSplit_outputEltWidth(unsigned bitWidth)
-{
-    if (bitWidth <= 8) {
-        return 1;
-    }
-    if (bitWidth <= 16) {
-        return 2;
-    }
-    if (bitWidth <= 32) {
-        return 4;
-    }
-    return 8;
-}
-
 int ZS_bitSplit_validateParams(
         const uint8_t* bitWidths,
         size_t nbWidths,
