+++ src/openzl/codecs/bitSplit/encode_bitSplit_kernel.h
@@ -6,19 +6,12 @@
 #include <stddef.h>
 #include <stdint.h>
 
+#include "openzl/codecs/bitSplit/bitSplit_common.h"
 #include "openzl/zl_portability.h"
 
 #if defined(__cplusplus)
 extern "C" {
 #endif
-
-/**
- * Determines the output element width for a given bit width.
- *
- * @param bitWidth Number of bits (1-64)
- * @return Output element width in bytes (1, 2, 4, or 8)
- */
-size_t ZS_bitSplit_outputEltWidth(unsigned bitWidth);
 
 /**
  * Encodes a single element by extracting bit ranges.
